changed code layout
