+++ cmd/agent/api/agent/agent.go
@@ -1,11 +1,19 @@
-package api
+package agent
 
 import (
 	"encoding/json"
 	"net/http"
 
+	"github.com/DataDog/datadog-agent/cmd/agent/stopper"
 	"github.com/DataDog/datadog-agent/pkg/version"
+	"github.com/gorilla/mux"
 )
+
+// SetupHandlers adds the specific handlers for /agent endpoints
+func SetupHandlers(r *mux.Router) {
+	r.HandleFunc("/version", getVersion).Methods("GET")
+	r.HandleFunc("/stop", stop).Methods("POST")
+}
 
 func getVersion(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
@@ -15,6 +23,6 @@ func getVersion(w http.ResponseWriter, r *http.Request) {
 }
 
 func stop(w http.ResponseWriter, r *http.Request) {
-	Stopper <- true
+	stopper.Stopper <- true
 	w.WriteHeader(202)
 }
