+++ cmd/agent/app/app.go
@@ -6,10 +6,12 @@ a public HTTP interface implementing several functionalities.
 package app
 
 import (
+	"fmt"
 	"path/filepath"
 
 	"github.com/DataDog/datadog-agent/pkg/collector/check"
 	"github.com/DataDog/datadog-agent/pkg/collector/scheduler"
+	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/kardianos/osext"
 	"github.com/spf13/cobra"
 )
@@ -36,3 +38,20 @@ monitoring and performance data.`,
 	// The Scheduler
 	_scheduler *scheduler.Scheduler
 )
+
+func setupConfig() {
+	// set the paths where a config file is expected
+	for _, path := range configPaths {
+		config.Datadog.AddConfigPath(path)
+	}
+
+	// load the configuration
+	err := config.Datadog.ReadInConfig()
+	if err != nil {
+		panic(fmt.Errorf("unable to load Datadog config file: %s", err))
+	}
+
+	// define defaults for the Agent
+	config.Datadog.SetDefault("cmd_sock", "/tmp/agent.sock")
+	config.Datadog.BindEnv("cmd_sock")
+}
