+++ cmd/agent/stopper/stopper.go
@@ -1,0 +1,4 @@
+package stopper
+
+// Stopper is the channel used by other packages to ask for stopping the agent
+var Stopper = make(chan bool)
